refactoring some more tests
